Finished jmh socket impl
